Extract footer component to @/components/ui
